move thumbnail creation into its own file
